break up some functions, move parse logic into own script
